requirements.txt
tk
Pillow
plyer
pystray
